requirements.txt
streamlit
requests
python-dotenv
pandas
flask
